refactor: Update package.json to use "postinstall" script for Prisma generation
The package.json file has been modified to replace the "prisma" and "generate" scripts with a single "postinstall" script. This script will automatically run the "prisma generate" command after the installation of dependencies. This change improves the convenience and maintainability of the project by simplifying the script configuration in package.json.