refactor: Add comment functionality to BlogPage component
The BlogPage component in the app/(main)/(route)/blog/[slug]/page.tsx file has been updated to include a comment section. This section allows users to leave comments on blog posts. The CommentInput and CommentItem components have been added to handle the input and display of comments respectively. This enhancement improves user engagement and encourages interaction on the website.