refactor: Update manage-tag component to include checkmark for edited tags
The manage-tag component in the app/(main)/(route)/dashboard/tag/_components/manage-tag.tsx file has been updated to include a checkmark icon for edited tags. When a tag is being edited, the checkmark icon is displayed instead of the pencil icon. This change improves the user experience by providing a clear visual indication of the editing state for tags.